Rename MBSP-TES3MP to TES3MP-MBSP
In consultation with others in the community, the convention is to name
projects `tes3mp-foobar`, not `foobar-tes3mp`. I was conflicted about
the name in this case: primarily, is this a TES3MP mod that happens to
be MBSP, or is it a version of MBSP that happens to be for TES3MP?

Given (1) that this mod was written from scratch, (2) that the mechanics
in this mod don't correspond 1-to-1 to MBSP, and (3) the conventions of
the TES3MP community, I'm going with TES3MP-MBSP, rather than the other
way around.

Glad to get that analysis paralysis moment out of the way, whew.